+++ cloudflare/schema_cloudflare_access_application.go
@@ -133,7 +133,7 @@ func resourceCloudflareAccessApplicationSchema() map[string]*schema.Schema {
 		"http_only_cookie_attribute": {
 			Type:     schema.TypeBool,
 			Optional: true,
-			Default:  false,
+			Computed: true,
 		},
 		"same_site_cookie_attribute": {
 			Type:     schema.TypeString,
@@ -152,6 +152,11 @@ func resourceCloudflareAccessApplicationSchema() map[string]*schema.Schema {
 			Type:     schema.TypeBool,
 			Optional: true,
 			Default:  true,
+		},
+		"service_auth_401_redirect": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			Default:  false,
 		},
 	}
 }